pyproject.toml
[project]
name = "cori"
version = "0.1.0"
description = "AI coauthor system for personalized cover letters"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "html2text>=2025.4.15",
    "httpx>=0.28.1",
    "pydantic-settings>=2.13.1",
    "uvicorn[standard]>=0.44.0",
]

[tool.ruff]
line-length = 100

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
